Add ability to disable monitoring subsystems
